common: Simplify user directory config via makedirs
makedirs acts like install -d, and can create missing intermediate
dirs for you.

This also introduces a pbwdir_path helper to address the common need
to get a subpath of the config directory. The initial user directory
splatting has been updated to adopt it, but not the rest of the code
in the file.